HW_06_v2
routes/api/users.js:
- created a new router file for user-related routes, importing the controller functions;

controllers/users.js:
- new folder/file created for better separation of tasks
